#1845, move QueryResponsePackets to common
